Changing the f.k mapping idea

+++ Benaa.Core/Entities/General/Rate.cs
@@ -13,7 +13,7 @@ namespace Benaa.Core.Entities.General
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-        public int Stars { get; set; }
+        public float Stars { get; set; }
         [Required]
         [ForeignKey(nameof(User))]
         public string StudentId { get; set; }
